Initialize flusher tid in hammer_flusher_create().
hmp->flusher.tid is usually updated after the flusher did its work.
If we have nothing to flush, hmp->flusher.tid will be zero.
This may result in a snapshot with TID = 0, if it is taken right
after mount before there is anything to flush. To prevent this,
we initialise hmp->flusher.tid with hmp->next_tid while mounting.